Add a check to see if the flash is writable
It seems like the Exx10 systems and newer don't set up SMI write
protections when the FDO is set by the EC, so disabling SMIs to bypass
them is unnecessary. The new logic should be able to differentiate
whether or not the SMM bypass is necessary and direct the user
appropriately.