e6400_flash_unlock.c: Clean up code
Remove the commented out read_ec_reg as it is not currently planned to
be utilized. Simply the function to send the EC FDO command, as the
other arguments to query and unset the FDO as well as the return value
are never used. The enums for the different arguments were removed,
though the values were moved to a comment for reference.