Rename to dell_flash_unlock
This is no longer limited to the E6400 so change it to a more general
name. While it currrently only seems to support Dell's business laptops
with MEC5035 derived ECs, it is possible other lineups will also work
with this, or this utility could be extended to support those if they
provide a similar FDO feature.